Fix MSVC compiler warnings
reg. 'return': conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/accessibility/accessibleitem.cpp
@@ -234,7 +234,7 @@ int AccessibleItem::accessibleSelectionCount() const
         return 0;
     }
 
-    return textCursor->selectedText().size();
+    return static_cast<int>(textCursor->selectedText().size());
 }
 
 int AccessibleItem::accessibleCursorPosition() const
